+++ CHANGELOG.md
@@ -6,6 +6,7 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 #### Feature
 - Automatically start next leg without needing to reload the page to improve load time, and get back to throwing quicker
 - Forward clients back to main page from match result after 2 minutes
+- New "Explore" tab on player statistics, to explore darts thrown
 
 #### Changed
 - Updated to use `Node.js v18`